refactor: move presign logic into attachment TRPC router
Replace the ad-hoc API route with a getUploadUrls mutation on the
existing attachmentRouter. orgProcedure already provides the verified
organizationId, eliminating the need to replicate auth and org checks.
Update usePresignedUpload to call api.attachment.getUploadUrls instead
of a raw fetch.
